requirement.txt
fastapi
uvicorn
pydantic
typing-extensions
langchain>=0.0.230
PyPDF2>=3.0.0
requests>=2.28.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.10
openai>=1.0.0
python-dotenv>=1.0.0
pymongo==4.6.0
pypdf